// src/app/common/components/Elements/AddressFieldsGroup01/types.ts
export interface AddressFieldsGroupProps {
  groupId?: string;
  id?: string[];
  name: string[];
  label?: string;
  syncLabel: string;

}

export interface ValidateFieldNames<T> {
  field0: T;
  field1: T;
  field3: T;
  field5: T;
  field6: T;
  field7: T;
  field8: T;
  field9: T;
  field10: T;
}